Switch API URL in production
Previously, all frontend fetch API calls were of the form '/users'
(i.e. they did not specify where to send the HTTP requests). In the
deployed (build / production) version of the frontend, the requests
were not linked to the deployed API server.
I used the NODE_ENV environment variable to change the URL all API
calls were made to depending on whether the frontend was being
deployed or not.